In-house Win32/WDK metadata generation
Replace the externally-sourced win32metadata `.winmd` with metadata generated
in-house directly from the Windows SDK and WDK headers.

- Add `windows-clang`, a new crate that scrapes C/C++ headers into RDL, split
  out from `windows-rdl` (which compiles RDL <-> `.winmd`).
- Add `tool_win32` and `tool_wdk` to produce `metadata/win32/*.rdl` and
  `metadata/wdk/*.rdl` and the corresponding `Windows.Win32.winmd` /
  `Windows.Wdk.winmd` bundled with `windows-bindgen`. WDK metadata is additive
  over Win32 and shares the flat `Windows.Win32` namespace.
- Faithful SAL/MIDL capture in the scraper: In/Out/opt, len/size counts,
  null-terminated string promotion (PCWSTR/PWSTR/...), COM out-ptr (`iid_is`),
  calling conventions, architecture attributes, and pseudo-attributes; callbacks
  route through the shared parameter parsing.
- Canonicalization of faithful header types to their metadata form centralized
  in `windows-clang` (numerics remap, scalar/pointer collapse, string aliases,
  interface-typedef omission).
- `windows-metadata` cut over to the in-house winmd (CLR element types authored
  in-source). `windows-version` cut over to WDK-scraped `RtlGetVersion`.
- `windows-bindgen` hardening validated by generating the whole Win32+WDK surface
  in `--minimal` and `--sys`: dedupe redundant `IID_<Interface>` constants, wide
  integer const literals, and `Default` on projected numerics in sys mode.
- Migrate reactor/canvas/webview/threading/window and related crates to the new
  metadata; simplify `windows-webview` string handling via `windows-strings`.
- The legacy monolithic `windows` / `windows-sys` crates remain generated from
  the reference `.winmd` via `tool_package` and are byte-identical.

+++ crates/libs/canvas/src/device.rs
@@ -41,7 +41,7 @@ impl GpuDevice {
                 std::ptr::null_mut(),
                 driver_type,
                 std::ptr::null_mut(),
-                D3D11_CREATE_DEVICE_BGRA_SUPPORT,
+                D3D11_CREATE_DEVICE_BGRA_SUPPORT as u32,
                 feature_levels.as_ptr(),
                 feature_levels.len() as u32,
                 D3D11_SDK_VERSION,
